backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# API clients
httpx==0.27.0
google-generativeai==0.8.0

# Settings and validation
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1

# CORS
starlette==0.38.0

# Rate limiting and caching
slowapi==0.1.9

# Async support
anyio==4.4.0
websockets==13.1

# Logging
structlog==24.4.0

# Production server
gunicorn==23.0.0
